+++ main.py
@@ -34,7 +34,7 @@ class TrackingUI:
         self.root.title("Marker Tracker - M3B Lab")
         self.root.iconphoto(False, tk.PhotoImage(file="ico/m3b_comp.png"))
 
-        self.preprocessVals = None
+        self.preprocess_vals = None
 
         self.setup_styles()
 
@@ -273,8 +273,12 @@ class TrackingUI:
         noisy_bg_radio.grid(row=4, column=0, pady=4)
         noisy_bg_radio = ttk.Radiobutton(self.area_frame, text="Harsh gradients\n(distinct darker edges or rings in video,\nwith lighter tracked subjects)", variable=self.preprocessing_choice_var, value=PreprocessingIssue.HARSH_GRADIENT.value, style='Regular.TRadiobutton')        
         noisy_bg_radio.grid(row=4, column=1, pady=4)
+        salt_pepper_radio = ttk.Radiobutton(self.area_frame, text="Salt+Pepper noise", variable=self.preprocessing_choice_var, value=PreprocessingIssue.SALT_PEPPER.value, style='Regular.TRadiobutton')        
+        salt_pepper_radio.grid(row=5, column=0, pady=(0,4))
+        custom_radio = ttk.Radiobutton(self.area_frame, text="Custom", variable=self.preprocessing_choice_var, value=PreprocessingIssue.CUSTOM.value, style='Regular.TRadiobutton')        
+        custom_radio.grid(row=5, column=1, pady=(0,4))
         no_processing_radio = ttk.Radiobutton(self.area_frame, text="No preprocessing", variable=self.preprocessing_choice_var, value=PreprocessingIssue.NONE.value, style='Regular.TRadiobutton')        
-        no_processing_radio.grid(row=5, column=0, columnspan=2, pady=(0,4))
+        no_processing_radio.grid(row=6, column=0, columnspan=2, pady=(0,4))
 
         # tracking operation buttons
         track_record_frame = tk.Frame(self.section1)
@@ -592,7 +596,7 @@ class TrackingUI:
 
     def frame_preprocessor(self):
         if self.video_path != "":
-            FramePreprocessor(self, self.video_path, self.preprocessVals)
+            FramePreprocessor(self, self.video_path, self.preprocess_vals)
         else:
             msg = "Select a video before opening the video preprocessing tool"
             error_popup(msg)
@@ -801,8 +805,6 @@ class TrackingUI:
                 msg = "ERROR: Please select a radio button for a tracking operation."
                 error_popup(msg)
             case TrackingOperation.MARKERS:
-                profiler = cProfile.Profile()
-                profiler.enable()
                 print("Beginning Marker Tracking Process...")
                 bbox_size = int(self.bbox_size_tracking_entry.get())
                 tracker_choice = TrackerChoice(self.tracker_choice_intvar.get())
@@ -816,7 +818,7 @@ class TrackingUI:
                     data_label_err_flag = self.check_data_label('output/Tracking_Output.csv', range_id)
 
                 if not data_label_err_flag:
-                    selected_markers, first_frame = tracking.select_markers(cap, bbox_size, self.frame_start, self.preprocessVals) # prompt to select markers
+                    selected_markers, first_frame = tracking.select_markers(cap, bbox_size, self.frame_start, self.preprocess_vals) # prompt to select markers
                     print(f"marker locs: {selected_markers}")
                     if not selected_markers.__contains__((-1,-1)): # select_markers returns list of -1 if selections cancelled
                         if use_multithread:
@@ -868,12 +870,7 @@ class TrackingUI:
                                     video_name,
                                     range_id
                                 )
-                        profiler.disable()
-                        stats = pstats.Stats(profiler)
-                        stats.sort_stats('cumulative').print_stats(10)
             case TrackingOperation.NECKING:
-                profiler = cProfile.Profile()
-                profiler.enable()
 
                 necking_method = NeckingPointMethod(self.necking_method_intvar.get())
                 binarize_intensity_thresh = int(self.binarize_intensity_thresh_entry.get())
@@ -953,14 +950,9 @@ class TrackingUI:
                     else:
                         msg = "Error: Please select a necking point method."
                         error_popup(msg)
-                        
-                    profiler.disable()
-                    stats = pstats.Stats(profiler)
-                    stats.sort_stats('cumulative').print_stats(10)
 
             case TrackingOperation.AREA:
-                profiler = cProfile.Profile()
-                profiler.enable()
+
                 # check if range_id already used
                 if file_mode == FileMode.APPEND: # only need to check prev ids if appending
                     data_label_err_flag = self.check_data_label('output/Surface_Area_Output.csv', range_id)
@@ -969,7 +961,7 @@ class TrackingUI:
                     bbox_size = int(self.bbox_size_area_entry.get())
                     distance_from_marker_thresh = int(self.distance_from_marker_thresh_entry.get())
                     preprocessing_need = PreprocessingIssue(self.preprocessing_choice_var.get())
-                    selected_markers, first_frame = tracking.select_markers(cap, bbox_size, self.frame_start, self.preprocessVals) # prompt to select markers
+                    selected_markers, first_frame = tracking.select_markers(cap, bbox_size, self.frame_start, self.preprocess_vals) # prompt to select markers
                     print(f"marker locs: {selected_markers}")
                     if not selected_markers.__contains__((-1,-1)): # select_markers returns list of -1 if selections cancelled
                         if use_multithread:    
@@ -1004,11 +996,8 @@ class TrackingUI:
                                 video_name,
                                 range_id,
                                 preprocessing_need,
-                                self.preprocessVals
+                                self.preprocess_vals
                             )
-                        profiler.disable()
-                        stats = pstats.Stats(profiler)
-                        stats.sort_stats('cumulative').print_stats(10)
     def get_file(self):
         """util function to prompt a file browser to select the video file that will be tracked
 
@@ -2155,7 +2144,7 @@ class FramePreprocessor:
             raise ValueError("Failed to read the video file.")
         
         # shrink and greyscale first frame, then copy for modded frame
-        self.first_frame, _ = tracking.scale_frame(cv2.cvtColor(self.first_frame, cv2.COLOR_BGR2GRAY), .5)
+        self.first_frame, _ = tracking.scale_frame(cv2.cvtColor(self.first_frame, cv2.COLOR_BGR2GRAY), .9)
         self.modded_frame = self.first_frame.copy()
 
         # create main window
@@ -2195,7 +2184,7 @@ class FramePreprocessor:
         self.create_checkbox_with_slider("Brightness", self.brightness_var, 3, -100, 100, self.basic_options_frame)
 
         # Create "Show Advanced" button
-        self.show_advanced_button = tk.Button(self.window, text="Show Advanced", command=self.toggle_advanced_options)
+        self.show_advanced_button = tk.Button(self.window, text="Show Custom Options - Advanced", command=self.toggle_advanced_options)
         self.show_advanced_button.grid(row=4, column=0, sticky=tk.W, padx=(110, 0))
 
         # smoothness options
@@ -2235,7 +2224,7 @@ class FramePreprocessor:
             self.load_preprocess_options(filename)
 
     def save_preprocess_options(self, filename):
-        preprocessVals = {
+        preprocess_vals = {
             "Blur/Sharpness": self.sliders["Blur/Sharpness"].get() if self.sharpness_var.get() else 0,
             "Contrast": self.sliders["Contrast"].get() if self.contrast_var.get() else 0,
             "Brightness": self.sliders["Brightness"].get() if self.brightness_var.get() else 0,
@@ -2243,24 +2232,24 @@ class FramePreprocessor:
             "Binarize": self.binarize_var.get()
         }
         with open(filename, 'w') as f:
-            json.dump(preprocessVals, f)
+            json.dump(preprocess_vals, f)
 
     def load_preprocess_options(self, filename):
         with open(filename, 'r') as f:
-            preprocessVals = json.load(f)
-        self.set_preprocess_values(preprocessVals)
+            preprocess_vals = json.load(f)
+        self.set_preprocess_values(preprocess_vals)
         self.update_preview()
 
-    def set_preprocess_values(self, preprocessVals):
-        self.sharpness_var.set(preprocessVals["Blur/Sharpness"] != 0)
-        self.contrast_var.set(preprocessVals["Contrast"] != 0)
-        self.brightness_var.set(preprocessVals["Brightness"] != 0)
-        self.smoothness_var.set(preprocessVals["Smoothness"] != 0)
-        self.binarize_var.set(preprocessVals["Binarize"])
-        self.sliders["Blur/Sharpness"].set(preprocessVals["Blur/Sharpness"])
-        self.sliders["Contrast"].set(preprocessVals["Contrast"])
-        self.sliders["Brightness"].set(preprocessVals["Brightness"])
-        self.sliders["Smoothness"].set(preprocessVals["Smoothness"])
+    def set_preprocess_values(self, preprocess_vals):
+        self.sharpness_var.set(preprocess_vals["Blur/Sharpness"] != 0)
+        self.contrast_var.set(preprocess_vals["Contrast"] != 0)
+        self.brightness_var.set(preprocess_vals["Brightness"] != 0)
+        self.smoothness_var.set(preprocess_vals["Smoothness"] != 0)
+        self.binarize_var.set(preprocess_vals["Binarize"])
+        self.sliders["Blur/Sharpness"].set(preprocess_vals["Blur/Sharpness"])
+        self.sliders["Contrast"].set(preprocess_vals["Contrast"])
+        self.sliders["Brightness"].set(preprocess_vals["Brightness"])
+        self.sliders["Smoothness"].set(preprocess_vals["Smoothness"])
 
     def create_checkbox_with_slider(self, text, variable, row, min_val, max_val, parent_frame):
         checkbox = ttk.Checkbutton(parent_frame, text=text, variable=variable)
@@ -2287,7 +2276,7 @@ class FramePreprocessor:
         max_label.grid_remove()
         value_label.grid_remove()
 
-        variable.trace_add("write", lambda *args, slider=slider, min_label=min_label, max_label=max_label, value_label=value_label: self.toggle_slider(slider, variable, min_label, max_label, value_label, parent_frame))
+        variable.trace_add("write", lambda *args: self.toggle_slider(slider, variable, min_label, max_label, value_label, parent_frame))
 
     def on_slider_change(self, value, value_label, variable):
         value_label.config(text=f"Value: {int(float(value))}")
@@ -2304,9 +2293,10 @@ class FramePreprocessor:
             min_label.grid_remove()
             max_label.grid_remove()
             value_label.grid_remove()
-
+            
         # Update the layout of the parent frame
         parent_frame.update_idletasks()
+        self.update_preview()
 
     def toggle_advanced_options(self):
         if self.advanced_options_visible:
@@ -2321,14 +2311,16 @@ class FramePreprocessor:
     def update_preview(self):
         if self.modded_frame is not None and self.modded_frame.size > 0:
             self.modded_frame = tracking.preprocess_frame(
-                self.first_frame, self.getPreprocessVals()
+                self.first_frame, self.get_preprocess_vals(), True
             )
             frame = cv2.cvtColor(self.modded_frame, cv2.COLOR_GRAY2RGB)
             imgtk = ImageTk.PhotoImage(image=Image.fromarray(frame))
             self.preview_label.imgtk = imgtk
             self.preview_label.configure(image=imgtk)
+        
+        self.window.lift()
 
-    def getPreprocessVals(self):
+    def get_preprocess_vals(self):
         returnDict = {
             "Blur/Sharpness": self.sliders["Blur/Sharpness"].get() if self.sharpness_var.get() else 0,
             "Contrast": self.sliders["Contrast"].get() if self.contrast_var.get() else 0,
@@ -2339,7 +2331,7 @@ class FramePreprocessor:
         return returnDict
 
     def on_close(self):
-        self.parent.preprocessVals = self.getPreprocessVals()
+        self.parent.preprocess_vals = self.get_preprocess_vals()
         self.cap.release()
         self.window.destroy()